refactor(tests): hoist duplicated wfl-binary spawn harness into tests/common
Twelve test files each re-implemented the same scaffolding for spawning the real
`wfl` CLI: locating the executable, writing a program into a fresh TempDir,
running it, and capturing combined stdout+stderr plus the exit code.

Hoisted into `tests/common/mod.rs`:

  wfl_exe()          -> env!("CARGO_BIN_EXE_wfl")      (Cargo-built test binary)
  wfl_release_exe()  -> target/release/wfl[.exe]       (separately built release)
  run_src()          -> write main.wfl, spawn, capture (String, Option<i32>)
  run_file_status()  -> run a named file with extra args

  13 files changed, 99 insertions(+), 181 deletions(-)   net -82 lines

The real CLI boundary is preserved everywhere -- only the wrapper moved. No test
was converted to an in-memory driver.

The two original `wfl_exe` variants resolve to DIFFERENT binaries: the
`CARGO_BIN_EXE_wfl` form gives the profile-matched test binary, while the
hardcoded path gives the separately built release binary. They are kept as two
distinct helpers rather than unified, and every call site retains the binary it
used before: seven on the debug binary (cli_help_version_flags,
package_protocol_removed, phase1_correctness_regression, transpiler_sunset,
github_issues_batch, property_index_access, wfl_version_builtin,
flush_action_backcompat) and four on the release binary
(docs_parser_and_include_fixes, subprocess_security, include_of_form_resolution,
load_module_undefined_hint).

`phase1_correctness_regression_test.rs` keeps its richer `run_files` driver with
the RUN_TIMEOUT wall-clock kill and pinned WFL_GLOBAL_CONFIG_PATH; only its
three-line `wfl_exe()` was hoisted. Migrating the plain `run_src` copies onto
that timeout-protected driver is a behavioral change, left as a follow-up.

Coverage evidence: `--list` inventories captured before and after for all twelve
targets are byte-identical, and the full-suite count is unchanged at 2096 passed.

Gates: cargo test --all 2096 passed / 0 failed / 27 ignored (159 suites); fmt
clean; clippy --all-targets --all-features -D warnings clean;
run_integration_tests.sh 136 passed / 0 failed / 24 skipped.
