+++ tests/include_of_form_resolution_test.rs
@@ -15,33 +15,10 @@
 //! still surfaced (fatal without includes, warning with includes).
 
 use std::fs;
-use std::process::Command;
 use tempfile::TempDir;
 
-fn wfl_exe() -> &'static str {
-    if cfg!(target_os = "windows") {
-        "target/release/wfl.exe"
-    } else {
-        "target/release/wfl"
-    }
-}
-
-/// Run a WFL file inside `dir` (so `include from` resolves sibling modules),
-/// returning (combined stdout+stderr, exit code).
-fn run_file_status(dir: &TempDir, name: &str, extra_args: &[&str]) -> (String, Option<i32>) {
-    let path = dir.path().join(name);
-    let output = Command::new(wfl_exe())
-        .args(extra_args)
-        .arg(&path)
-        .output()
-        .expect("Failed to execute WFL");
-    let combined = format!(
-        "{}{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    (combined, output.status.code())
-}
+mod common;
+use common::run_file_status;
 
 /// Write a `mod.wfl` exposing a one-arg `greet` action next to a `main.wfl`
 /// whose body is `main_body`, then run `main.wfl`.
