+++ tests/package_protocol_removed_test.rs
@@ -20,9 +20,8 @@ use std::path::Path;
 use std::process::Command;
 use tempfile::TempDir;
 
-fn wfl_exe() -> &'static str {
-    env!("CARGO_BIN_EXE_wfl")
-}
+mod common;
+use common::wfl_exe;
 
 /// Run `wfl <args...>` with `dir` as the working directory.
 /// Returns (stdout, stderr, exit code).
